requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# LangChain Core (Install these FIRST)
langchain-core==0.1.52
langchain-community==0.0.38
langchain==0.1.20

# Google Gemini
google-generativeai==0.3.2
langchain-google-genai==0.0.11

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Vector Store & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Medical NLP - Compatible versions
spacy==3.5.3
scispacy==0.5.1

# Utilities
pandas==2.0.3
numpy==1.24.3

# Frontend
streamlit==1.28.0

# Additional dependencies
typing-extensions==4.7.1
requests==2.31.0